Fix bug in diagnosing multiple null semantics: t+
When a nulled symbol had more than one semantics, and none was distinquised
by being the semantics of an explicity empty rule, the documentation stated that
Marpa would fail with a message.  In fact, Marpa was often simply going ahead with one semantics,
picked arbitrarily.  This commit changes Marpa::R2 to behave as documented.