Android requires some workarounds before calling configure
Three env variables need to be defined on android, to both
work around bugs in the OS and in the configure script.

First, TMPDIR must be set due to a bug in android's sh:
http://stackoverflow.com/questions/15283220/android-shell-eof

Second, CONFIG_SHELL and PATH_SEPARATOR need to be set
since Android's sh is in /system/bin/sh, and several spots
in configure assume that /bin/sh will be available.